code refactor merge (#4)
* Update README.md

* Update readme (#2)

* changes for prod branch

* Update README.md

* Update README.md

* Update README.md

* refactoring to add more endpoints

* refactoring backend

* dev environment changes for frontend and refactor

* refactoring reservations

* booking flow process has started

* pickup and dropoff times now show for reservations

* reservation crud added

* car won't display if dropoff isn't before a pickup

* sql statement to get all car's reservation ranges

* typo with sql statement

* more sql tuning

* reservation dates display in edit

* car form controls are dynamically built

* location search "starting with"

* locations now auto complete

* Can't delete or update parent row: reservation_id

* delete works on reservations through a join table

* car deletes now - idk what fixed it

* changed all refrences from booking to reservation

* changed all booking refrences to reservation

* changed book to reserve

* changed book to reserve

* delete and reserve view complete

* added some logging and test data

* Refactor (#3)

* locations now auto complete

* Can't delete or update parent row: reservation_id

* delete works on reservations through a join table

* car deletes now - idk what fixed it

* changed all refrences from booking to reservation

* changed all booking refrences to reservation

* changed book to reserve

* changed book to reserve

* delete and reserve view complete

* added some logging and test data

* Update README.md

* reorganized layout to add utility folder

* Reorganized (#5)

* locations now auto complete

* Can't delete or update parent row: reservation_id

* delete works on reservations through a join table

* car deletes now - idk what fixed it

* changed all refrences from booking to reservation

* changed all booking refrences to reservation

* changed book to reserve

* changed book to reserve

* delete and reserve view complete

* added some logging and test data

* reorganized layout to add utility folder

* trying to resolve merge conflicts

* cleaning up files from the prod merge conflict

* cleaning up from the marge conflict (#6)

* locations now auto complete

* Can't delete or update parent row: reservation_id

* delete works on reservations through a join table

* car deletes now - idk what fixed it

* changed all refrences from booking to reservation

* changed all booking refrences to reservation

* changed book to reserve

* changed book to reserve

* delete and reserve view complete

* added some logging and test data

* reorganized layout to add utility folder

* trying to resolve merge conflicts

* cleaning up files from the prod merge conflict

* Forgot to delete the file locally

+++ frontend/src/app/app.component.ts
@@ -1,12 +1,15 @@
-import { AfterViewInit, Component, OnInit, ViewChild } from '@angular/core';
+import { AfterContentInit, AfterViewInit, Component, OnInit, ViewChild } from '@angular/core';
 import { AddCarFormComponent } from './add-car-form/car-form.component';
 import { CarDiscoverResultsTableDataSource } from '../car-discover-results-table.datasource';
 import { carDiscoverHTTPService } from '../car-discover-http.service';
-import { FormGroup } from '@angular/forms';
+import { locationHttpService } from 'src/services/location/location.service';
+import { AbstractControl, FormBuilder, FormControl, FormGroup, Validators } from '@angular/forms';
 import { ResultsTableComponent } from './table/results-table/results-table.component';
 import { CardiscoverFormComponent } from './cardiscover-form/cardiscover-form.component';
 import { transition } from '@angular/animations';
-
+import { ReservationHTTPService } from 'src/services/reservation/reservation.service';
+import { of } from 'rxjs';
+import { MatSnackBar } from '@angular/material/snack-bar';
 @Component({
   selector: 'app-root',
   templateUrl: './app.component.html',
@@ -15,7 +18,7 @@ import { transition } from '@angular/animations';
 })
 
 
-export class AppComponent {
+export class AppComponent  {
 
 
   rowsClicked: any[] = [];
@@ -34,6 +37,10 @@ export class AppComponent {
   editView = false;
   // additional DO location results table 
   showDoLocationResults = false;
+  showReservationResults = false;
+  hideReservationsButtonText = false;
+  hideResultsTable = false;
+  showReservations = false;
   noInputView = false;
   filteredCarResultsView = false;
   responseCarId = 0;
@@ -41,16 +48,23 @@ export class AppComponent {
   showTableChooseCarText = false;
   chosenRowModifyPointer = 0;
   previousChosenRowModifyPointer = 0;
+  showCarReservedText = false;
   carResultsDataSource = new CarDiscoverResultsTableDataSource();
   doLocationResultsDataSource = new CarDiscoverResultsTableDataSource();
+  reservationResultsDataSource = new CarDiscoverResultsTableDataSource();
+  @ViewChild('reservationsResults') reservationResultsTableComponent!: ResultsTableComponent;
   @ViewChild('carResults') carResultsTableComponent!: ResultsTableComponent;
   @ViewChild('doLocationResults') doLocationResults!: ResultsTableComponent;
   @ViewChild('cardiscoverForm') cardiscoverForm!: CardiscoverFormComponent;
   @ViewChild('carForm') addCarFormComponentChild!: AddCarFormComponent;
 
-
+  temporaryCarForm!:FormGroup;
   constructor(
+    private formBuilder: FormBuilder, 
     private carDiscover: carDiscoverHTTPService,
+    private locationService: locationHttpService,
+    private reservationService: ReservationHTTPService,
+    private snackBar: MatSnackBar
     ) {}
 
   /* 
@@ -58,6 +72,7 @@ export class AppComponent {
   * Functions for toggling views
   *
   */
+
 
   filterForSelected(){
     if(this.showTable){
@@ -78,6 +93,27 @@ export class AppComponent {
     }
   }
 
+  filterForPickUpLocationName(event:any){
+    let currentString = event.target.value;
+    if(currentString.length >= 3){
+      if(event.keycode !== 8){
+        this.locationService.getLocationNames(currentString).subscribe((responseData) => {
+          this.cardiscoverForm.filteredPickupOptions = of(responseData)
+        })
+      }
+    }
+  }
+
+  filterForDropoffLocationName(event:any){
+    let currentString = event.target.value;
+    if(currentString.length >= 3){
+      if(event.keycode !== 8){
+        this.locationService.getLocationNames(currentString).subscribe((responseData) => {
+          this.cardiscoverForm.filteredDropoffOptions = of(responseData)
+        })
+      }
+    }
+  }
 
   clearSelected(){
     // reset the clicked rows to not selected state and clear the array
@@ -108,9 +144,8 @@ export class AppComponent {
       this.noInputView = false;
       this.insertNewCarTextHeader = true;
       this.rowsClicked = [];
-      if(this.addCarView === false){
-        this.showCarDiscoverForm = true;
-      }
+      this.addCarView = false;
+      this.showCarDiscoverForm = true;
     } 
     else {
       this.noInputView = false;
@@ -170,7 +205,7 @@ export class AppComponent {
 
   getAvaliableCars(emittedData: any){
     const formData = emittedData
-    this.carDiscover.getCars(formData.doLocation, formData.puLocation, formData.doDate, formData.puDate).subscribe((response: any) => {
+    this.locationService.getCars(formData.doLocation, formData.puLocation, formData.doDate, formData.puDate).subscribe((response: any) => {
       // if nothing gets returned
       if( Object.keys(response).length === 0){
         if(this.noInputView) this.noInputView = false
@@ -221,18 +256,85 @@ export class AppComponent {
     // }
   }
 
+  resetView(){
 
-  
-  deleteAllCars(){
-    this.carDiscover.deleteAllCars().subscribe()
+    this.noResult = false;  
+    this.showCarDiscoverForm = true;
+    this.hideReservationsButtonText = false;
+    this.showReservationResults = false
+    this.showTable = false
+    this.showCarReservedText = false;
+    this.clearSelected()
   }
 
-  bookCar(){
-    const formData = this.cardiscoverForm;
-    this.carDiscover.bookCar(formData.doLocation, formData.puLocation, 
-      formData.doDate, formData.puDate, this.rowsClicked[0]).subscribe((response: any) => {
-      console.log(formData.doLocation, formData.puLocation, 
-        formData.doDate, formData.puDate, this.rowsClicked[0])})
+  getAllReservations(){
+    if(this.showReservationResults){
+      this.resetView()
+      return
+    }
+    this.noInputView = false;
+    this.showReservationResults = true;
+    // this.editView = true;
+    this.reservationService.getAllReservations().subscribe((response: any) => {
+      if( Object.keys(response).length === 0){
+        if(this.noResult === true){
+          document.body.querySelector(".shake")?.getAnimations()[0].play();
+        }
+        this.noResult = true;  
+        if(this.showReservationResults){
+          this.showReservationResults = false;
+        }
+        return;
+      }
+      // Show table results with all cars response
+      this.reservationResultsDataSource.setData([...response]);
+      // reservationResultsDataSource = new CarDiscoverResultsTableDataSource();
+      // @ViewChild('reservationsResults') reservationResultsTableComponent!: ResultsTableComponent;
+      this.reservationResultsTableComponent.displayedColumns = Object.keys(response[0]);
+      // Ensure the car's ID is always at the front if present
+      let idPos = this.reservationResultsTableComponent.displayedColumns.indexOf("id")
+      if(idPos !== -1){
+        this.reservationResultsTableComponent.displayedColumns.splice(idPos, 1);
+        this.reservationResultsTableComponent.displayedColumns.unshift("id");
+      }
+      // This will show only the car results table with controls,  
+      // showDoLocationResults needs to be set to true to also show doLocation results table
+      this.reservationResultsTableComponent.showTable = true;
+      this.noResult = false;  
+      this.showCarDiscoverForm = false;
+      this.hideReservationsButtonText = true;
+      this.showTable = true;
+      if(this.hideResultsTable){
+        this.carResultsDataSource.setData([])
+        this.showTable = false;
+      }
+    })
+
+  }
+
+  
+  
+  deleteAllCars(){
+    if(this.rowsClicked.length >= 2){
+      this.carDiscover.deleteAllCars().subscribe()
+    }
+  }
+
+  reserveCar(){
+    let formData = this.cardiscoverForm.carDiscoverForm.value;
+    console.log("this is the payload", formData.doLocation, 
+    formData.puLocation, formData.doDate, formData.puDate, this.rowsClicked[0].id)
+
+    this.reservationService.reserveCar(formData.doLocation, formData.puLocation, 
+      formData.doDate, formData.puDate, this.rowsClicked[0].id).subscribe((response: any) => {
+        this.resetView()
+        this.clearSelected();
+        this.hideResultsTable = true;
+        this.getAllReservations();
+        this.snackBar.open("Car has been reserved","Dismiss")._dismissAfter(6_000);
+    })
+    
+
   }
 
   modify(){
@@ -251,26 +353,39 @@ export class AppComponent {
                 return;
               }
               if(this.addCarFormComponentChild){
+                if(this.temporaryCarForm){
+                  this.addCarFormComponentChild.carForm = this.temporaryCarForm;
+                }
                 this.addCarFormComponentChild.carForm.patchValue({
-                  id: responseData.id,
+                  id: Number.parseInt(responseData.id),
                   brand: responseData.brand,
-                  dropoffTime: responseData.dropoffTime,
+                  dropoffTime: new Date(Date.parse(responseData.dropoffTime)),
                   location: responseData.location,
                   model: responseData.model,
-                  pickupTime: responseData.pickupTime,
+                  pickupTime: new Date(Date.parse(responseData.pickupTime)),
                   rate: responseData.rate,
                   size: responseData.size,
                   supplier: responseData.supplier
                 });
+                console.table(responseData);
+                if('reservation_id' in responseData){
+                  this.addCarFormComponentChild.carForm.patchValue({
+                    reservation_id: Number.parseInt(responseData.reservation_id),
+                  });
+                  this.addCarFormComponentChild.reservation_id = responseData.reservation_id;
+                } else {
+                  this.addCarFormComponentChild.carForm.removeControl('reservation_id');
+                  this.addCarFormComponentChild.reservation_id = -1;
+                }
+                this.addCarFormComponentChild.location = responseData.location;
               }else {
                 console.error("NOT RENDERED YET??!!!")
 
               }
-              console.log(this.addCarFormComponentChild.carForm.value.rate)
+              console.log(this.addCarFormComponentChild.carForm.value)
               this.responseCarId = parseInt(responseData.id)
             }
           );
-
           this.addCarView = true;
           this.editCarTextHeader = true;
           this.insertNewCarTextHeader = false;
@@ -282,9 +397,71 @@ export class AppComponent {
         console.error(this.addCarFormComponentChild)
       }
     }
+    if(this.showReservationResults){
+      if(this.addCarFormComponentChild){
+        if(this.rowsClicked.length > 0){
+          // Will allow using Next & Previous button
+          this.previousChosenRowModifyPointer = this.chosenRowModifyPointer
+          let editReservation = this.rowsClicked[this.chosenRowModifyPointer];
+          this.addCarFormComponentChild.editView = true;
+          // this.addCarFormComponentChild.clearSelectedView = true;
+          this.reservationService.getReservation(editReservation.reservation_id).subscribe(
+            (responseData:any) => {
+              if( Object.keys(responseData).length === 0){
+                console.error("Couldn't get reservation ID", editReservation)
+                return;
+              }
+              if(this.addCarFormComponentChild){
+              console.log("this is the response data", responseData)
+                this.addCarFormComponentChild.carForm.patchValue({
+                  location: null,
+                  id: Number.parseInt(responseData.id),
+                  reservation_id: Number.parseInt(responseData.reservation_id),
+                  dropoffTime: new Date(Date.parse(responseData.dropoffTime)),
+                  pickupTime: new Date(Date.parse(responseData.pickupTime)),
+                });
+              }else {
+                console.error("NOT RENDERED YET??!!!")
+
+              }
+              this.temporaryCarForm = this.addCarFormComponentChild.carForm;
+              console.log(this.addCarFormComponentChild.carForm.value)
+              
+              this.responseCarId = parseInt(responseData.id)
+            }
+          );
+          // this.addCarFormComponentChild.carForm.removeControl('location');
+          this.addCarFormComponentChild.location = '';
+
+          this.addCarView = true;
+          // this.editCarTextHeader = true;
+          this.insertNewCarTextHeader = false;
+          window.scroll({top: 0, behavior: "smooth"})
+        }
+      }
+      else {
+        console.error("Component hasn't rendered or is null!")
+        console.error(this.addCarFormComponentChild)
+      }
+    }
   }
 
-  deleteRows(){}
+  deleteRows(event: any){
+    if(this.showReservationResults){
+        this.rowsClicked.forEach(rowElement => {
+          this.reservationService.deleteReservation(rowElement.reservation_id).subscribe();
+        });
+        this.getAllReservations();
+
+    }
+    if(this.editView){
+      this.rowsClicked.forEach(rowElement => {
+        this.carDiscover.deleteCar(rowElement.id).subscribe();
+      });
+      this.toggleEditView();
+    }
+    this.snackBar.open("Items have been deleted", "Dismiss")._dismissAfter(6_000);
+  }
